refactor(intents): rename factmanagement package and fix lint issues
- Rename package from fact_management to factmanagement (Go convention)
- Fix context.Context parameter order in NewIntentContext
- Refactor handlers to reduce cognitive complexity
- Convert single-case switches to if statements
- Fix IsNewFact bug where it was checked after CancelEdit cleared it
- Add nolint comments for intentional error ignores in tests

AI-Generated-By: OpenCode (Claude Sonnet 4)
