+++ internal/cli/intents/factmanagement/fact_management_suite_test.go
@@ -1,4 +1,4 @@
-package fact_management
+package factmanagement
 
 import (
 	"testing"
