+++ internal/cli/intents/factmanagement/result_test.go
@@ -1,22 +1,22 @@
-package fact_management_test
+package factmanagement_test
 
 import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/baphled/kariya/internal/cli/intents/fact_management"
+	"github.com/baphled/kariya/internal/cli/intents/factmanagement"
 	"github.com/baphled/kariya/internal/domain/career"
 )
 
 var _ = Describe("Result", func() {
 	Describe("Construction", func() {
 		It("should create an empty result", func() {
-			result := &fact_management.Result{}
+			result := &factmanagement.Result{}
 			Expect(result).NotTo(BeNil())
 		})
 
 		It("should store action", func() {
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Action: "created",
 			}
 			Expect(result.Action).To(Equal("created"))
@@ -24,7 +24,7 @@ var _ = Describe("Result", func() {
 
 		It("should store fact", func() {
 			fact := &career.Fact{ID: "fact-1", Text: "Test fact"}
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Fact: fact,
 			}
 			Expect(result.Fact).To(Equal(fact))
@@ -35,21 +35,21 @@ var _ = Describe("Result", func() {
 				{ID: "fact-1", Text: "Fact 1"},
 				{ID: "fact-2", Text: "Fact 2"},
 			}
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Facts: facts,
 			}
 			Expect(result.Facts).To(HaveLen(2))
 		})
 
 		It("should store message", func() {
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Message: "Operation completed successfully",
 			}
 			Expect(result.Message).To(Equal("Operation completed successfully"))
 		})
 
 		It("should store scroll position", func() {
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				ScrollPosition: 42,
 			}
 			Expect(result.ScrollPosition).To(Equal(42))
@@ -58,7 +58,7 @@ var _ = Describe("Result", func() {
 
 	Describe("Action Types", func() {
 		It("should support created action", func() {
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Action:  "created",
 				Message: "Fact created successfully",
 			}
@@ -66,7 +66,7 @@ var _ = Describe("Result", func() {
 		})
 
 		It("should support updated action", func() {
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Action:  "updated",
 				Message: "Fact updated successfully",
 			}
@@ -74,7 +74,7 @@ var _ = Describe("Result", func() {
 		})
 
 		It("should support deleted action", func() {
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Action:  "deleted",
 				Message: "Fact deleted successfully",
 			}
@@ -82,7 +82,7 @@ var _ = Describe("Result", func() {
 		})
 
 		It("should support none action", func() {
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Action: "none",
 			}
 			Expect(result.Action).To(Equal("none"))
@@ -93,7 +93,7 @@ var _ = Describe("Result", func() {
 		It("should contain all fields for a complete operation", func() {
 			fact := &career.Fact{ID: "fact-1", Text: "Test fact"}
 			facts := []*career.Fact{fact}
-			result := &fact_management.Result{
+			result := &factmanagement.Result{
 				Action:         "created",
 				Fact:           fact,
 				Facts:          facts,
